Restructure service
