+++ tests/test_service.py
@@ -7,14 +7,11 @@ from unittest.mock import patch
 import pandas as pd
 import pytz
 
-# Import the objects we want to test from the pysportbot.service module
-from pysportbot.service import (
-    calculate_class_day,
-    calculate_next_execution,
-    run_service,
-    validate_config,
-    validate_config_and_activities,
-)
+from pysportbot.service.config_validator import validate_config, validate_config_and_activities
+
+# 1) Import each function from its exact module path:
+from pysportbot.service.scheduling import calculate_class_day, calculate_next_execution
+from pysportbot.service.service import run_service
 from pysportbot.utils.errors import ErrorMessages
 
 
@@ -102,7 +99,7 @@ class TestService(unittest.TestCase):
     # 2. Tests for validating activities against SportBot
     # ------------------------------------------------------------------------
 
-    @patch("pysportbot.service.SportBot")  # Correct import path
+    @patch("pysportbot.service.config_validator.SportBot")  # Correct import path to where SportBot is used
     def test_validate_config_and_activities_unknown_activity(self, mock_sportbot_class):
         """
         Test that if the config specifies an activity not present in SportBot activities,
@@ -138,7 +135,7 @@ class TestService(unittest.TestCase):
     # 3. Tests for the utility functions that calculate execution dates
     # ------------------------------------------------------------------------
 
-    @patch("pysportbot.service.datetime", wraps=datetime)  # Use the real datetime for unmocked methods
+    @patch("pysportbot.service.scheduling.datetime", wraps=datetime)  # Patch datetime where it's actually used
     def test_calculate_next_execution(self, mock_datetime):
         """
         If today is Wednesday 2024-01-10, then 'Friday 07:30:00' should be 2 days ahead (2024-01-12).
@@ -156,7 +153,7 @@ class TestService(unittest.TestCase):
         # Assert
         self.assertEqual(result, expected)
 
-    @patch("pysportbot.service.datetime")  # Correct import path
+    @patch("pysportbot.service.scheduling.datetime")  # Correct import path
     def test_calculate_class_day(self, mock_datetime):
         """
         If today is Wednesday 2024-01-10, the next Monday is 2024-01-15 (5 days later).
@@ -173,10 +170,10 @@ class TestService(unittest.TestCase):
     # 4. Tests for the main booking logic (weekly vs. one-off)
     # ------------------------------------------------------------------------
 
-    @patch("pysportbot.service.datetime", wraps=datetime)
-    @patch("pysportbot.service.SportBot")
-    @patch("pysportbot.service.schedule.run_pending")
-    @patch("pysportbot.service.time.sleep", return_value=None)
+    @patch("pysportbot.service.scheduling.datetime", wraps=datetime)
+    @patch("pysportbot.service.service.SportBot")
+    @patch("pysportbot.service.service.schedule.run_pending")
+    @patch("pysportbot.service.service.time.sleep", return_value=None)
     def test_weekly_scheduling(self, mock_sleep, mock_run_pending, mock_sportbot_class, mock_datetime):
         """
         Test that run_service schedules a weekly job and (theoretically) calls schedule.run_pending() in a loop.
@@ -204,7 +201,7 @@ class TestService(unittest.TestCase):
         }
 
         # Ensure schedule.jobs is patched to avoid infinite loops
-        with patch("pysportbot.service.schedule.jobs", new=[]):
+        with patch("pysportbot.service.service.schedule.jobs", new=[]):
             run_service(
                 config,
                 offset_seconds=0,
@@ -216,10 +213,10 @@ class TestService(unittest.TestCase):
         mock_bot_instance.login.assert_called_with("test@example.com", "password")
         self.assertTrue(mock_bot_instance.login.called, "Expected login to be called")
 
-    @patch("pysportbot.service.datetime")
-    @patch("pysportbot.service.SportBot")
-    @patch("pysportbot.service.schedule.run_pending")
-    @patch("pysportbot.service.time.sleep", return_value=None)
+    @patch("pysportbot.service.scheduling.datetime")
+    @patch("pysportbot.service.service.SportBot")
+    @patch("pysportbot.service.service.schedule.run_pending")
+    @patch("pysportbot.service.service.time.sleep", return_value=None)
     def test_now_scheduling(self, mock_sleep, mock_run_pending, mock_sportbot_class, mock_datetime):
         """
         Test that a one-off booking with "now" triggers an immediate booking
@@ -236,22 +233,20 @@ class TestService(unittest.TestCase):
         # Booking date should be 2024-01-08 (Monday)
         mock_bot_instance.daily_slots.return_value = pd.DataFrame({"start_timestamp": ["2024-01-08 18:00:00"]})
 
-        config = {
-            "email": "test@example.com",
-            "password": "password",
-            "classes": [
-                {
-                    "activity": "Yoga",
-                    "class_day": "Monday",
-                    "class_time": "18:00:00",
-                    "booking_execution": "now",
-                    "weekly": False,
-                }
-            ],
-        }
-
         run_service(
-            config,
+            {
+                "email": "test@example.com",
+                "password": "password",
+                "classes": [
+                    {
+                        "activity": "Yoga",
+                        "class_day": "Monday",
+                        "class_time": "18:00:00",
+                        "booking_execution": "now",
+                        "weekly": False,
+                    }
+                ],
+            },
             offset_seconds=0,
             retry_attempts=1,
             retry_delay_minutes=0,
@@ -269,10 +264,10 @@ class TestService(unittest.TestCase):
     # 5. Tests for error handling (no matching slots, already booked, retries)
     # ------------------------------------------------------------------------
 
-    @patch("pysportbot.service.SportBot")
-    @patch("pysportbot.service.schedule.run_pending")
-    @patch("pysportbot.service.time.sleep", return_value=None)
-    def test_no_matching_slots_error(self, mock_sleep, mock_run_pending, mock_sportbot_class):
+    @patch("pysportbot.service.booking.time.sleep", return_value=None)
+    @patch("pysportbot.service.service.schedule.run_pending")
+    @patch("pysportbot.service.service.SportBot")
+    def test_no_matching_slots_error(self, mock_sportbot_class, mock_run_pending, mock_sleep):
         """
         If daily_slots() returns no row matching class_time, we expect a ValueError.
         However, run_service handles it internally and logs an error instead of raising.
@@ -297,7 +292,7 @@ class TestService(unittest.TestCase):
         }
 
         # We need to mock datetime to return a fixed date for calculate_class_day
-        with patch("pysportbot.service.datetime") as mock_datetime:
+        with patch("pysportbot.service.scheduling.datetime") as mock_datetime:
             tz = pytz.timezone("Europe/Madrid")
             fixed_now = tz.localize(datetime(2024, 1, 7, 12, 0, 0))  # Sunday
             mock_datetime.now.return_value = fixed_now
@@ -305,12 +300,16 @@ class TestService(unittest.TestCase):
 
             with self.assertLogs("pysportbot.service", level="WARNING") as cm:
                 run_service(
-                    config, offset_seconds=0, retry_attempts=1, retry_delay_minutes=0, time_zone="Europe/Madrid"
+                    config,
+                    offset_seconds=0,
+                    retry_attempts=1,
+                    retry_delay_minutes=0,
+                    time_zone="Europe/Madrid",
                 )
 
         # Expected log messages:
-        # WARNING  pysportbot.service:service.py:155 Attempt 1 failed for Yoga: No matching slots available for Yoga at 18:00:00 on 2024-01-08
-        # ERROR    pysportbot.service:service.py:167 Failed to book Yoga after 1 attempts.
+        # WARNING  pysportbot.service: Attempt 1 failed for Yoga: ...
+        # ERROR    pysportbot.service: Failed to book Yoga after 1 attempts.
 
         expected_warning = ErrorMessages.no_matching_slots_for_time("Yoga", "18:00:00", "2024-01-08")
         self.assertIn(expected_warning, cm.output[0])
@@ -318,10 +317,10 @@ class TestService(unittest.TestCase):
         expected_error = "Failed to book Yoga after 1 attempts."
         self.assertIn(expected_error, cm.output[1])
 
-    @patch("pysportbot.service.SportBot")  # Correct import path
-    @patch("pysportbot.service.schedule.run_pending")  # Correct import path
-    @patch("pysportbot.service.time.sleep", return_value=None)  # Correct import path
-    def test_slot_already_booked_error(self, mock_sleep, mock_run_pending, mock_sportbot_class):
+    @patch("pysportbot.service.booking.time.sleep", return_value=None)
+    @patch("pysportbot.service.service.schedule.run_pending")
+    @patch("pysportbot.service.service.SportBot")
+    def test_slot_already_booked_error(self, mock_sportbot_class, mock_run_pending, mock_sleep):
         """
         If bot.book(...) raises a slot_already_booked error, the code should skip further retries.
         """
@@ -348,7 +347,7 @@ class TestService(unittest.TestCase):
         }
 
         # Mock datetime to set the booking date to 2024-01-08
-        with patch("pysportbot.service.datetime") as mock_datetime:
+        with patch("pysportbot.service.scheduling.datetime") as mock_datetime:
             tz = pytz.timezone("Europe/Madrid")
             fixed_now = tz.localize(datetime(2024, 1, 7, 12, 0, 0))  # Sunday
             mock_datetime.now.return_value = fixed_now
@@ -359,20 +358,16 @@ class TestService(unittest.TestCase):
                     config, offset_seconds=0, retry_attempts=2, retry_delay_minutes=1, time_zone="Europe/Madrid"
                 )
 
-        # Expected log messages:
-        # WARNING  pysportbot.service:service.py:155 Attempt 1 failed for Yoga: The slot is already booked.
-        # ERROR    pysportbot.service:service.py:167 Failed to book Yoga after 2 attempts.
-
         expected_warning = already_booked_msg
         self.assertIn(expected_warning, cm.output[0])
 
         # Ensure that book was called only once
         mock_bot_instance.book.assert_called_once()
 
-    @patch("pysportbot.service.SportBot")
-    @patch("pysportbot.service.schedule.run_pending")
-    @patch("pysportbot.service.time.sleep", return_value=None)
-    def test_retry_mechanism(self, mock_sleep, mock_run_pending, mock_sportbot_class):
+    @patch("pysportbot.service.booking.time.sleep", return_value=None)
+    @patch("pysportbot.service.service.schedule.run_pending")
+    @patch("pysportbot.service.service.SportBot")
+    def test_retry_mechanism(self, mock_sportbot_class, mock_run_pending, mock_sleep):
         """
         If booking fails for a reason other than 'already booked',
         the service should retry up to 'retry_attempts'.
@@ -402,7 +397,7 @@ class TestService(unittest.TestCase):
         }
 
         # Mock datetime to set the booking date to 2024-01-08
-        with patch("pysportbot.service.datetime") as mock_datetime:
+        with patch("pysportbot.service.scheduling.datetime") as mock_datetime:
             tz = pytz.timezone("Europe/Madrid")
             fixed_now = tz.localize(datetime(2024, 1, 7, 12, 0, 0))  # Sunday
             mock_datetime.now.return_value = fixed_now
@@ -414,9 +409,9 @@ class TestService(unittest.TestCase):
                 )
 
         # Expected log messages:
-        # WARNING  pysportbot.service:service.py:155 Attempt 1 failed for Yoga: Some random booking error
-        # WARNING  pysportbot.service:service.py:155 Attempt 2 failed for Yoga: Some random booking error
-        # ERROR    pysportbot.service:service.py:167 Failed to book Yoga after 2 attempts.
+        # WARNING  pysportbot.service: Attempt 1 failed for Yoga: Some random booking error
+        # WARNING  pysportbot.service: Attempt 2 failed for Yoga: Some random booking error
+        # ERROR    pysportbot.service: Failed to book Yoga after 2 attempts.
 
         expected_warning_1 = "Attempt 1 failed for Yoga: Some random booking error"
         expected_warning_2 = "Attempt 2 failed for Yoga: Some random booking error"
@@ -431,7 +426,7 @@ class TestService(unittest.TestCase):
         self.assertEqual(mock_bot_instance.book.call_count, 2)
         self.assertTrue(mock_sleep.called, "Expected to sleep between retries")
 
-    @patch("pysportbot.service.datetime", wraps=datetime)
+    @patch("pysportbot.service.scheduling.datetime", wraps=datetime)
     def test_calculate_next_execution_time_already_passed(self, mock_datetime):
         """
         If today is Monday 2024-12-30 and the time is 07:00:00,
@@ -449,8 +444,8 @@ class TestService(unittest.TestCase):
         expected = tz.localize(datetime(2025, 1, 6, 6, 30, 0))
         self.assertEqual(result, expected)
 
-    @patch("pysportbot.service.time.sleep", return_value=None)
-    @patch("pysportbot.service.SportBot")
+    @patch("pysportbot.service.booking.time.sleep", return_value=None)
+    @patch("pysportbot.service.service.SportBot")
     def test_slot_matching(self, mock_sportbot_class, mock_sleep):
         """
         Ensure the service correctly matches a slot with a specific time.
